refactor(gateway): 优化链路驱动和测试组件配置
- 提取 modelMapping 变量避免重复调用 resolveAccountModelMapping 方法
- 为 OpenAI 推理字段标准化添加上游端点路径转换功能
- 重构认证过期同步组件支持测试注入时间间隔参数
- 简化调度测试中的重试等待逻辑并提高测试效率
- 更新调度计划规范化测试的基线数据结构和验证规则
- 优化批量更新操作的目标组装和版本冲突检测机制
- 移除不可达的 JSON 解码防御代码分支以简化实现

+++ backend-go/projects/gateway/internal/accounts/w14l_tx_sweep_test.go
@@ -72,7 +72,7 @@ func w14lSequences() map[string]w14lSequence {
 					AvailabilitySchedulePresent: true,
 					AvailabilitySchedule: map[string]any{
 						"enabled": true, "timezone": "UTC", "mode": "allow_windows",
-						"windows": []any{map[string]any{"daysOfWeek": []any{1, 2}, "start": "01:00", "end": "02:00"}},
+						"windows": []any{map[string]any{"daysOfWeek": []any{float64(1), float64(2)}, "start": "01:00", "end": "02:00"}},
 					},
 				}, f.scope())
 				return err
